change power result to long

+++ TP1/exo1.cpp
@@ -4,13 +4,12 @@
 
 #define return_and_display(result) return _.store(result);
 
-int power(int value, int n)
+long power(long value, long n)
 {
     Context _("power", value, n); // do not care about this, it allow the display of call stack
 
     // your code
-
-
+    // return_and_display shows the result in the window and returns it
     return_and_display(1);
 }
